💄 Improve contrast of modal text
Modal text was using #ccc which wasn't very good on the dark background, so it's been bumped to #eee to give a contrast ratio of  15.74 instead of 11.37.

Modal footer text has also been switched from the default theme's #666 to #999 on both themes, bumping the contrast ratio from a dangerous 3.18 to 6.41.