✨ Automatic light/dark theme detection (#7)
We now use CSS L5 Media Queries to automatically apply the theme we think the user would like for the forum. This information is obtained from the browser, which obtains it from the system's selected theme (light or dark). If the media query is not supported on the device, the light theme will remain in use, like normal.

If the user's device has a dark theme selected, we will enable dark mode automatically, but also display a Modal, allowing the user to stay in dark mode or switch to light mode. If they choose to switch back to light mode, a flag will be set on their account to make sure that they don't automatically switch again. If they choose to remain on dark mode, the website will continue auto-selecting their theme per device but will not display the modal again. Furthermore, there is a note at the bottom of the Modal, informing the user how to change their mind later on.

If the user ever manually switched between light and dark mode, the site will not auto-select the theme for that user again. This does not apply if toggling OLED Dark Mode, however.

This feature still needs soem feedback. Any of the Modal text can be edited easily in the locale YML file. The JS behind the feature isn't beautiful but it was the easiest way that 16 year old me could implement it without rewriting the whole extension.

Fixes #7